more scripts fixes

+++ WorkloadManagementSystem/scripts/dirac-admin-show-task-queues.py
@@ -9,9 +9,9 @@
 from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
+
 __RCSID__ = "$Id$"
 
-from past.builtins import long
 import sys
 
 from DIRAC import S_OK, gLogger
@@ -35,7 +35,7 @@ taskQueueID = 0
 
 def setTaskQueueID(optVal):
   global taskQueueID
-  taskQueueID = long(optVal)
+  taskQueueID = int(optVal)
   return S_OK()
